+++ lib/src/search/local_search.dart
@@ -1,31 +1,23 @@
+import '../models/combo_box_item_model.dart';
 import '../models/typedefs.dart';
 import 'search_engine.dart';
 
-class LocalSearch<T>
-    implements SearchEngine<T> {
-  LocalSearch({
-    required this.items,
-    required this.displayText,
-  });
+class LocalSearch implements SearchEngine {
+  LocalSearch({required this.items, required this.displayText});
 
-  final List<T> items;
+  final List<ComboBoxItemModel> items;
 
-  final DisplayText<T> displayText;
+  final DisplayText<ComboBoxItemModel> displayText;
 
   @override
-  Future<List<T>> search(
-      String keyword,
-      ) async {
+  Future<List<ComboBoxItemModel>> search(String keyword) async {
     if (keyword.isEmpty) {
       return items;
     }
 
     return items.where((e) {
-      return displayText(e)
-          .toLowerCase()
-          .contains(
-        keyword.toLowerCase(),
-      );
+      return displayText(e).toLowerCase().contains(keyword.toLowerCase());
     }).toList();
   }
+
 }