+++ lib/src/search/search_engine.dart
@@ -1,5 +1,5 @@
-abstract class SearchEngine<T> {
-  Future<List<T>> search(
-      String keyword,
-      );
+import '../models/combo_box_item_model.dart';
+
+abstract class SearchEngine {
+  Future<List<ComboBoxItemModel>> search(String keyword);
 }