+++ lib/src/widgets/combo_box_item.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
-class ComboBoxItem<T>
-    extends StatelessWidget {
+import '../models/combo_box_item_model.dart';
+
+class ComboBoxItem extends StatelessWidget {
   const ComboBoxItem({
     super.key,
     required this.item,
@@ -9,9 +10,13 @@ class ComboBoxItem<T>
     required this.selected,
     required this.highlighted,
     required this.onTap,
+    required this.textStyle,
+    required this.textStyleSelected,
+    required this.bgColorSelected,
+    required this.heightItem,
   });
 
-  final T item;
+  final ComboBoxItemModel item;
 
   final String title;
 
@@ -21,32 +26,33 @@ class ComboBoxItem<T>
 
   final VoidCallback onTap;
 
-  @override
-  Widget build(
-      BuildContext context,
-      ) {
-    Color? background;
+  final TextStyle? textStyle;
 
+  final TextStyle? textStyleSelected;
+
+  final Color? bgColorSelected;
+
+  final double heightItem;
+
+
+
+  @override
+  Widget build(BuildContext context) {
+    Color? background;
     if (highlighted) {
-      background =
-          Colors.blue.shade50;
+      background = Colors.blue.shade50;
     } else if (selected) {
-      background =
-          Colors.grey.shade100;
+      background = Colors.grey.shade100;
     }
 
     return InkWell(
       onTap: onTap,
       child: Container(
-        height: 48,
-        color: background,
-        padding:
-        const EdgeInsets.symmetric(
-          horizontal: 12,
-        ),
-        alignment:
-        Alignment.centerLeft,
-        child: Text(title),
+        height: heightItem,
+        color: bgColorSelected??background,
+        padding: const EdgeInsets.symmetric(horizontal: 12),
+        alignment: Alignment.centerLeft,
+        child: Text(title, style: selected? textStyleSelected : textStyle),
       ),
     );
   }