+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-## 1.0.0
-
+## 1.0.1
+- Fix overlay position.
+- Fix search item.
+- Improve keyboard navigation.